Refactor onion payload validation (#631)
We rely less on exceptions and more on explicit types to encode errors.
This allows simplifying the `IncomingPaymentPacket.decrypt` methods by
separating concerns.

+++ src/commonTest/kotlin/fr/acinq/lightning/wire/FailureMessageTestsCommon.kt
@@ -51,7 +51,7 @@ class FailureMessageTestsCommon : LightningTestSuite() {
             FinalIncorrectHtlcAmount(123456.msat),
             ChannelDisabled(0, 1, channelUpdate),
             ExpiryTooFar,
-            InvalidOnionPayload(561.toULong(), 1105),
+            InvalidOnionPayload(561, 1105),
             PaymentTimeout,
             TrampolineFeeInsufficient,
             TrampolineExpiryTooSoon
